atlas_api/atlas_api_app/pyproject.toml
[project]
name = "veritas-atlas"
version = "1.2.0"
description = "Veritas Atlas API - Block 12 auth hardening scaffold"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "httpx>=0.27.0",
  "PyJWT[crypto]>=2.9.0",
  "sqlalchemy>=2.0.0",
  "psycopg2-binary>=2.9.0",
  "alembic>=1.13.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "celery[redis]>=5.3.0",
  "redis>=5.0.0",
  "boto3>=1.34.0",
]

[project.scripts]
atlas-api = "app.cli:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.23.8",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
